% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapDisparosVistaxG.R
\name{xGgraph}
\alias{xGgraph}
\title{Esta funcion crea una visualizacion a partir de datos de entrada de disparos introducidos de manera manual.}
\usage{
xGgraph()
}
\value{
El mapa de tiros y acumulado en un partido para los equipos con el xG calculado
}
\description{
Esta funcion crea una visualizacion a partir de datos de entrada de disparos introducidos de manera manual.
}
\examples{
xGgraph()
}
